LLM/Langchain/requirements.txt
langchain
ipykernel
python-dotenv
langchain-google-genai
langchain-openai
streamlit